requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
email-validator==2.2.0
python-dotenv==1.0.0
PyJWT==2.10.1